# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NuweFarma API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/activity-logs": {
            "get": {
                "tags": [
                    "Logs de Actividad"
                ],
                "summary": "Listar logs de actividad",
                "operationId": "38649a59c116e933561db69c56b6dbe2",
                "parameters": [
                    {
                        "name": "accion",
                        "in": "query",
                        "description": "Filtrar por acción",
                        "schema": {
                            "type": "string",
                            "example": "login"
                        }
                    },
                    {
                        "name": "modulo",
                        "in": "query",
                        "description": "Filtrar por módulo",
                        "schema": {
                            "type": "string",
                            "example": "usuarios"
                        }
                    },
                    {
                        "name": "usuario_id",
                        "in": "query",
                        "description": "Filtrar por ID de usuario",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "fecha_desde",
                        "in": "query",
                        "description": "Fecha desde (Y-m-d)",
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-01-01"
                        }
                    },
                    {
                        "name": "fecha_hasta",
                        "in": "query",
                        "description": "Fecha hasta (Y-m-d)",
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-01-31"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Registros por página",
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de logs de actividad",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "usuario_id": {
                                                        "type": "string"
                                                    },
                                                    "accion": {
                                                        "type": "string"
                                                    },
                                                    "modulo": {
                                                        "type": "string"
                                                    },
                                                    "descripcion": {
                                                        "type": "string"
                                                    },
                                                    "ip_address": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    },
                                                    "usuario": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/activity-logs/{id}": {
            "get": {
                "tags": [
                    "Logs de Actividad"
                ],
                "summary": "Ver detalles de un log específico",
                "operationId": "cbe7e05644606372130eed5f15b26f93",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del log",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "usuario_id": {
                                                    "type": "string"
                                                },
                                                "accion": {
                                                    "type": "string"
                                                },
                                                "modulo": {
                                                    "type": "string"
                                                },
                                                "registro_id": {
                                                    "type": "string"
                                                },
                                                "descripcion": {
                                                    "type": "string"
                                                },
                                                "datos_anteriores": {
                                                    "type": "object"
                                                },
                                                "datos_nuevos": {
                                                    "type": "object"
                                                },
                                                "ip_address": {
                                                    "type": "string"
                                                },
                                                "user_agent": {
                                                    "type": "string"
                                                },
                                                "created_at": {
                                                    "type": "string"
                                                },
                                                "usuario": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Log no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/activity-logs/me": {
            "get": {
                "tags": [
                    "Logs de Actividad"
                ],
                "summary": "Ver mis propios logs de actividad",
                "operationId": "8d6a310af66f07d725736df6a0b262c7",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Registros por página",
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mis logs de actividad",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}